requirements.txt
asgiref==3.5.2
autopep8==1.7.0
coverage==6.4.4
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
importlib-metadata==4.12.0
Markdown==3.4.1
Pillow==9.2.0
pycodestyle==2.9.1
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
zipp==3.8.1
